+++ src/LastAssignment/common.h
@@ -1,8 +1,8 @@
-#ifndef _COMMON_H_
-#define _COMMON_H_
+#ifndef COMMON_H_
+#define COMMON_H_
 
 static const char* SOURCE_DIRECTORY = "C:/graphics/src/LastAssignment/";
 static const char* SHADER_DIRECTORY = "C:/graphics/src/LastAssignment/shaders/";
 static const char* DATA_DIRECTORY = "C:/graphics/src/LastAssignment/data/";
 
-#endif  // _COMMON_H_
+#endif  // COMMON_H_
